+++ api/CHANGELOG.md
@@ -15,7 +15,7 @@
 - Update the "Show & allow using universal patches" setting, when toggled off, to hide apps on the app selection screen that don't use any patches (so ones that only use universal patches)
 - Converted the existing components of the app that use old file pickers to the new one along with saving and overwrite warning logic
 - Removed the "Show other versions" button on the select an app screen for apps that support all versions https://github.com/Jman-Github/Universal-ReVanced-Manager/issues/87
-
+- Patch options menus now follow the same design as the rest of the patch selection page https://github.com/Jman-Github/Universal-ReVanced-Manager/issues/84
 
 # Bug fixes